radv: Add new VK_MESA_query_timestamp extension to radv driver [v2]
This extension adds a single function to query the current GPU
timestamp, just like glGetInteger64v(GL_TIMESTAMP, &timestamp). This
function is needed to complete the implementation of
GOOGLE_display_timing, which needs to be able to correlate GPU and CPU
timestamps.

Signed-off-by: Keith Packard <keithp@keithp.com>

v2:	Update to track extension API that now returns both device and
	surface timestamps.